request/curl: extract curl-easy specific code to file curl/easy.rs
